CI process corrections

+++ src/services/embedding_service.rs
@@ -158,6 +158,7 @@ impl EmbeddingService {
 
     /// Generate embedding using Ollama
     async fn generate_embedding(&self, content: &str) -> Result<Vec<f32>, EmbeddingError> {
+        // NEW API: EmbeddingsInput::Single for a single string
         let input = EmbeddingsInput::Single(content.to_string());
         let request = GenerateEmbeddingsRequest::new("nomic-embed-text:latest".to_string(), input);
 
@@ -167,17 +168,18 @@ impl EmbeddingService {
             return Err(EmbeddingError::NoEmbeddings);
         }
 
-        // Extract the first (and only) embedding with proper casting
+        // The embeddings might be Vec<Vec<f64>> for batch, or Vec<f64> for single
+        // Try to extract the first embedding if it's nested
         let embedding: Vec<f32> = match response.embeddings.len() {
             0 => return Err(EmbeddingError::NoEmbeddings),
-            1 => response.embeddings[0].iter().map(|&v| v as f32).collect(),
+            1 => response.embeddings[0].iter().map(|&v| v).collect(),
             _ => response
                 .embeddings
                 .into_iter()
                 .next()
                 .unwrap()
                 .into_iter()
-                .map(|v| v as f32)
+                .map(|v| v)
                 .collect(),
         };
